+++ src/app/api/stores/route.ts
@@ -1,6 +1,4 @@
 import { NextResponse } from "next/server";
-import type { NextApiRequest, NextApiResponse } from "next";
-import { StoreApiResponse, StoreType } from "@/interface";
 import prisma from "@/db";
 import axios from "axios";
